new commit

+++ include/Globals.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 namespace nn::global {
-using ValueType = double;
+using ValueType = float;
 using ParamMetrix = std::vector<ValueType>;
 } // namespace nn::global
